2requirements2.txt
aiohttp==2.0.7
argcomplete==0.8.1
async-timeout==1.2.0
autobahn==0.17.1
auxlib==0.0.42
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
boto==2.46.1
bz2file==0.98
certifi==2017.4.17
chardet==3.0.4
chardet2==2.0.3
cloudpickle==0.2.1
colorama==0.2.5
command-not-found==0.3
constantly==15.1.0
cssselect==1.0.1
cycler==0.10.0
cymem==1.31.2
decorator==4.0.10
defer==1.0.6
defusedxml==0.5.0
dill==0.2.6
dj-database-url==0.4.1
Django==1.11.3
django-admin-oauth2==1.0.2
django-allauth==0.32.0
django-braces==1.11.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.1
django-rest-framework-social-oauth2==1.0.8
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.6.3
docker-py==1.10.3
docker-pycreds==0.2.1
docopt==0.6.2
entrypoints==0.2.2
enum34==1.1.6
falcon==1.1.0
fastzbarlight==0.0.14
feedfinder2==0.0.4
feedparser==5.2.1
gensim==2.1.0
go-vncdriver==0.4.19
gunicorn==19.6.0
gym==0.7.2
html5lib==0.999
httplib2==0.8
idna==2.5
incremental==16.10.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
jieba3k==0.35.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
Keras==0.3.1
language-check==1.0
language-selector==0.1
lda==1.0.4
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.3
menulibre==2.0.4
mistune==0.7.3
Morfessor==2.0.1
motor==1.1
mugshot==0.2.5
multidict==2.1.4
murmurhash==0.26.4
nbconvert==4.2.0
nbformat==4.1.0
newspaper3k==0.2.2
nltk==3.2.4
notebook==4.2.3
num2words==0.5.4
numpy==1.11.2
oauthlib==2.0.2
olefile==0.44
onboard==1.0.1
oneconf==0.3.7.14.4.1
OpenCC==0.2
packaging==16.8
pandas==0.19.1
parsedatetime==2.1
pathlib==1.0.1
pbr==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pipreqs==0.4.9
piston-mini-client==0.7.5
plac==0.9.6
polyglot==16.7.4
preshed==0.46.4
progressbar==2.3
prompt-toolkit==1.0.7
protobuf==3.1.0.post1
psutil==1.2.1
psycopg2==2.6.2
ptyprocess==0.5.1
pycld2==0.31
pycrypto==2.6.1
pycurl==7.19.3
pyenchant==1.6.8
pyglet==1.2.4
Pygments==2.1.3
pygobject==3.12.0
PyICU==1.9.5
PyJWT==1.5.2
pymongo==3.4.0
python-apt===0.9.3.5ubuntu2
python-dateutil==2.6.0
python-gnupg==0.3.6
python-logstash==0.4.6
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.2
pyxDamerauLevenshtein==1.4.1
pyxdg==0.25
PyYAML==3.12
pyzmq==15.4.0
qtconsole==4.2.1
requests==2.18.1
requests-file==1.4.2
requests-oauthlib==0.8.0
requirements==0.1
scikit-learn==0.18.1
scipy==0.18.1
semver==2.7.2
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
smart-open==1.5.2
social-auth-app-django==1.2.0
social-auth-core==1.4.0
software-center-aptd-plugins==0.0.0
spacy==1.2.0
sputnik==0.9.3
tensorflow==0.12.1
terminado==0.6
Theano==0.9.0
thinc==5.0.8
tldextract==2.1.0
tornado==4.4.1
traitlets==4.3.0
Twisted==16.6.0
txaio==2.6.0
Ubuntu-Make==16.4
ufw===0.34-rc-0ubuntu2
ujson==1.35
unattended-upgrades==0.1
universe==0.21.2
urllib3==1.21.1
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.40.0
whitenoise==2.0.6
widgetsnbextension==1.2.6
word2veckeras==0.0.5.2
xgboost==0.6a2
yarg==0.1.9
yarl==0.10.0
zope.interface==4.3.3
